functionality related to lifecycle transactions in a separate module
