+++ apps/indexer/lib/indexer/fetcher/arbitrum/workers/new_l1_executions.ex
@@ -24,8 +24,8 @@ defmodule Indexer.Fetcher.Arbitrum.Workers.NewL1Executions do
 
   import Explorer.Helper, only: [decode_data: 2]
 
-  alias Indexer.Fetcher.Arbitrum.Utils.Db
   alias Indexer.Fetcher.Arbitrum.Utils.Db.Messages, as: DbMessages
+  alias Indexer.Fetcher.Arbitrum.Utils.Db.ParentChainTransactions, as: DbParentChainTransactions
   alias Indexer.Fetcher.Arbitrum.Utils.Helper, as: ArbitrumHelper
   alias Indexer.Fetcher.Arbitrum.Utils.Rpc
   alias Indexer.Helper, as: IndexerHelper
@@ -300,7 +300,7 @@ defmodule Indexer.Fetcher.Arbitrum.Workers.NewL1Executions do
     lifecycle_transactions =
       basic_lifecycle_transactions
       |> ArbitrumHelper.extend_lifecycle_transactions_with_ts_and_status(blocks_to_ts, track_finalization?)
-      |> Db.get_indices_for_l1_transactions()
+      |> DbParentChainTransactions.get_indices_for_l1_transactions()
 
     executions =
       basics_executions
@@ -413,7 +413,7 @@ defmodule Indexer.Fetcher.Arbitrum.Workers.NewL1Executions do
 
       messages_map
       |> Map.keys()
-      |> Db.Messages.l1_executions()
+      |> DbMessages.l1_executions()
       |> Enum.map(fn execution ->
         messages_map
         |> Map.get(execution.message_id)
